labページにフッターを追加
