+++ src/pages/gallery/layout.tsx
@@ -1,9 +1,9 @@
-import React, { CSSProperties, ReactNode } from "react"
+import React, { CSSProperties } from "react"
 import ReactDOM from "react-dom"
-import { Link, ThemeProvider, Typography } from "@material-ui/core"
+import { ThemeProvider } from "@material-ui/core"
 import { defaultTheme } from "../../react-components/common/default_theme"
-import { webConstants } from "../../react-components/common/web_constants"
 import { Display } from "../../react-components/art/display"
+import { Footer } from "../../react-components/common/footer"
 
 const path = (relativePath: string): string => `..${relativePath}`
 const Title = () => {
@@ -17,36 +17,6 @@ const Title = () => {
   return (
     <div>
       <img src={path("/resources/title.svg")} style={style} /> {/* Dunno how to show Futura as text */}
-    </div>
-  )
-}
-
-const Footer = () => {
-  const footerStyle: CSSProperties = {
-    paddingTop: "5rem",
-    paddingBottom: "5rem",
-    backgroundColor: defaultTheme.palette.grey["400"],
-  }
-  const linkStyle: CSSProperties = {
-    display: "table",
-    margin: "auto",
-  }
-  const createSeparator = (): ReactNode => <Typography display="inline">  /  </Typography>
-
-  // FixMe: variant="body1"はTheme適用のためだが正しい操作ではない気がする
-  const createLink = (link: string, title: string): ReactNode => <Link variant="body1" href={link}>{title}</Link>
-  const createExternalLink = (link: string, title: string): ReactNode =>
-    <Link variant="body1" target="_blank" rel="noopener" href={link}>{title}</Link>
-
-  return (
-    <div style={footerStyle}>
-      <div style={linkStyle}>
-        {createLink(path("/"), "Home")}
-        {createSeparator()}
-        {createExternalLink(webConstants.twitterProfileUrl, "Twitter")}
-        {createSeparator()}
-        {createExternalLink(webConstants.instagramProfileUrl, "Instagram")}
-      </div>
     </div>
   )
 }
@@ -74,7 +44,7 @@ const App = () => {
         backgroundColor={backgroundListColor2}
         link={path("/pages/drawer.html?system.run=0&system.auto_download=0")}
       />
-      <Footer />
+      <Footer homePath={path("/")} />
     </ThemeProvider>
   )
 }
